Move Layout to pages folder too
